Remove `hard_strict` option from daterange._Getter
Removes `hard_strict` option from `daterange._Getter.get_start` and
clients. Also amends affected tests.

`hard_strict` was  provided to allow for raising an error, even when
strict passed as False, if period defined from a duration and 'start'
value that was before the left limit. However, this treatment was
incoherent with fact that could return partial prices by passing strict
False and defining end within available range and a duration that
resulted in the evaluated start falling before the left limit (i.e. the
same situation evaluated from other end point).

A review of code, tests and documentation found no justification
for forcing the raising of the error when strict False, which in
turn left the hard_strict redundant.