Update dependencies, bump max supported Python to 3.13
Includes updating yahooquery to v2.4.0 and removing patches
that had been introduced pending yq maintenance (and which
was included to 2.4.0).

Also fixes `TestGetterDaily.do_bounds_from_start` to recognise
that for monthly interval the end will be 'today' if, before rolling
back to prior month end, the end would be 'today' or later.