Fix `PricesBase.prices_for_symbols`
Reworks `PricesBase.prices_for_symbols` to not require a
deepcopy of the `data.Data` object. The previous approach
was failing after upgrading yahooquery (to 2.4.0). This failure
seems to have been a result of an object being introduced that
could not be pickled in the deepcopy process (it was something
on the YahooPrices object itself which was being included in
the deepcopy as a result of the `YahooPrices._request_data`
method being stored on `data.Data`. In short, the previous
deepcopy approach was overkill in any event. New approach
simply instantiates a new copy  of `data.Data`, for each base
interval, and sets the `_table` and `_ranges` attributes to copies
of the existing.